+++ tests/providers/google/cloud/hooks/test_cloud_sql.py
@@ -42,6 +42,10 @@ from tests.providers.google.cloud.utils.base_gcp_mock import (
 
 
 class TestGcpSqlHookDefaultProjectId:
+    def test_delegate_to_runtime_error(self):
+        with pytest.raises(RuntimeError):
+            CloudSQLHook(api_version="v1", gcp_conn_id="test", delegate_to="delegate_to")
+
     def setup_method(self):
         with mock.patch(
             "airflow.providers.google.common.hooks.base_google.GoogleBaseHook.__init__",