+++ tests/providers/google/cloud/hooks/test_gcs.py
@@ -129,6 +129,10 @@ class TestFallbackObjectUrlToObjectNameAndBucketName:
 
 
 class TestGCSHook:
+    def test_delegate_to_runtime_error(self):
+        with pytest.raises(RuntimeError):
+            gcs.GCSHook(api_version="v1", gcp_conn_id="GCP_CONN_ID", delegate_to="delegate_to")
+
     def setup_method(self):
         with mock.patch(
             GCS_STRING.format("GoogleBaseHook.__init__"),
@@ -753,6 +757,36 @@ class TestGCSHook:
                 mock.call().__exit__(None, None, None),
             ]
         )
+
+    @pytest.mark.parametrize(
+        "prefix, result",
+        (
+            (
+                "prefix",
+                [mock.call(delimiter=",", prefix="prefix", versions=None, max_results=None, page_token=None)],
+            ),
+            (
+                ["prefix", "prefix_2"],
+                [
+                    mock.call(
+                        delimiter=",", prefix="prefix", versions=None, max_results=None, page_token=None
+                    ),
+                    mock.call(
+                        delimiter=",", prefix="prefix_2", versions=None, max_results=None, page_token=None
+                    ),
+                ],
+            ),
+        ),
+    )
+    @mock.patch(GCS_STRING.format("GCSHook.get_conn"))
+    def test_list(self, mock_service, prefix, result):
+        mock_service.return_value.bucket.return_value.list_blobs.return_value.next_page_token = None
+        self.gcs_hook.list(
+            bucket_name="test_bucket",
+            prefix=prefix,
+            delimiter=",",
+        )
+        assert mock_service.return_value.bucket.return_value.list_blobs.call_args_list == result
 
     @mock.patch(GCS_STRING.format("GCSHook.get_conn"))
     def test_list_by_timespans(self, mock_service):