+++ tests/providers/google/cloud/hooks/test_pubsub.py
@@ -53,7 +53,6 @@ LABELS = {"airflow-version": "v" + version.replace(".", "-").replace("+", "-")}
 def mock_init(
     self,
     gcp_conn_id,
-    delegate_to=None,
     impersonation_chain=None,
 ):
     pass
@@ -73,6 +72,10 @@ def _generate_messages(count) -> list[ReceivedMessage]:
 
 
 class TestPubSubHook:
+    def test_delegate_to_runtime_error(self):
+        with pytest.raises(RuntimeError):
+            PubSubHook(gcp_conn_id="GCP_CONN_ID", delegate_to="delegate_to")
+
     def setup_method(self):
         with mock.patch(BASE_STRING.format("GoogleBaseHook.__init__"), new=mock_init):
             self.pubsub_hook = PubSubHook(gcp_conn_id="test")